project_management/requirements.txt
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.9
drf-yasg==1.21.7
